Add Vector2 class (including limited function cascading)
By function cascading, I am referring to the ability to specify a list 
of functions in `functionNames`. Cascading: if parsing for one element 
fails, then we try the next. In this case a function of three arguments 
would have too many arguments for `Vector2`, and it would cascade to 
`Vector3`. Limited: probably doesn't work for all failure conditions.